+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+* Added an easter egg
+* Updated Tagalog translation
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.11.8 BETA**