+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Test and beta releases will have experimental functions enabled by default.
 
 * Fixed defaced values for get network operator/sim info
 * Fixed phone state listener cast errors
+* Updated Lithuanian translation
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)