+++ CHANGELOG.md
@@ -16,11 +16,13 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.11.10 TEST**
+
 * Fixed defaced values for get network operator/sim info
 * Fixed phone state listener cast errors
 * Updated Lithuanian translation
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 1.11.9 TEST**